+++ src/material/standard/standard-material.ts
@@ -286,10 +286,6 @@ export class StandardMaterial extends Material {
     shader.uniforms.u_BaseColorFactor = this._baseColor
     shader.uniforms.u_ModelMatrix = mesh.worldTransform.array
     shader.uniforms.u_NormalMatrix = mesh.transform.normalTransform.array
-    if (this._shadowCastingLight) {
-      shader.uniforms.u_ShadowSampler = this._shadowCastingLight.shadowTexture
-      shader.uniforms.u_LightViewProjectionMatrix = this._shadowCastingLight.lightViewProjection
-    }
     if (this._alphaMode === StandardMaterialAlphaMode.mask) {
       shader.uniforms.u_AlphaCutoff = this.alphaCutoff
     }
@@ -320,6 +316,11 @@ export class StandardMaterial extends Material {
       shader.uniforms[`u_Lights[${i}].intensity`] = light.intensity
       shader.uniforms[`u_Lights[${i}].innerConeCos`] = Math.cos(light.innerConeAngle * DEG_TO_RAD)
       shader.uniforms[`u_Lights[${i}].outerConeCos`] = Math.cos(light.outerConeAngle * DEG_TO_RAD)
+    }
+    if (this._shadowCastingLight) {
+      shader.uniforms.u_ShadowSampler = this._shadowCastingLight.shadowTexture
+      shader.uniforms.u_LightViewProjectionMatrix = this._shadowCastingLight.lightViewProjection
+      shader.uniforms.u_ShadowLightIndex = lightingEnvironment.lights.indexOf(this._shadowCastingLight.light)
     }
     let imageBasedLighting = lightingEnvironment.imageBasedLighting
     if (imageBasedLighting?.valid) {